Add automatic generation of RSA key if the keyfile doesn't exist
This can make it easier to deploy SSH honeypots across large networks.  Ideally, there should be command-line parameters to control number of bits for the RSA key and whatnot, and I'd be happy to add these types of options if people make such requests, but for now I'm purposely keeping this simple to ease merging with other pull requests that add other interesting and useful command-line options.